Move BehaviorMode to gui.utils
